Updated script which generates the data for the open-trivia Java API

+++ database/convert-json-to-java-api-format.sh
@@ -1,23 +1,35 @@
 #!/bin/bash
 
-json_file_temp=temp.json
+json_file_temp1=temp1.json
+json_file_temp2=temp2.json
 json_file=open-trivia-java-api.json
 
-grep category open-trivia.json > $json_file_temp
+[ -e $json_file_temp1 ] && rm -f $json_file_temp1
+[ -e $json_file_temp2 ] && rm -f $json_file_temp2
 
-touch $json_file
-cat /dev/null > $json_file
+grep category open-trivia.json > $json_file_temp1
 
 i=1
-total=$(wc -l $json_file_temp)
-
+total=$(cat $json_file_temp1 | wc -l)
 while read line; do
-  echo "Processing line $i from $total"
-  echo $(echo $line | sed "s/{/{ \"id\": $i, /g" | sed "s/}/},/g") >> $json_file
-  i=$((i+1));
-done < $json_file_temp
+  echo "Processing line $i from $total..."
+  echo $(echo $line | sed "s/^{/{ \"id\": $i, /g" | sed "s/}/},/g") >> $json_file_temp2
+  i=$((i + 1));
+done < $json_file_temp1
 
-rm -f $json_file_temp
 
+# Remove last ,
+number_of_head_lines=$(cat $json_file_temp2 | wc -l)
+number_of_head_lines=$(($number_of_head_lines - 1))
+
+head -n $number_of_head_lines $json_file_temp2 > $json_file
+tail -n 1 $json_file_temp2 | sed 's/},/}/g' >> $json_file
+
+
+# Append [ to the beginning and ] at the end
 sed -i '1s/^/[\n/' $json_file
 sed -i '$a]' $json_file
+
+
+# Remove temp files
+rm -f $json_file_temp1 $json_file_temp2
